+++ src/com/ansorgit/plugins/bash/editor/codecompletion/VariableNameCompletionProvider.java
@@ -87,26 +87,26 @@ class VariableNameCompletionProvider extends AbstractBashCompletionProvider {
     private int addCollectedVariables(PsiElement element, CompletionResultSet result, BashVarCollectorProcessor processor) {
         PsiTreeUtil.treeWalkUp(processor, element, BashPsiUtils.findFileContext(element), ResolveState.initial());
 
-        Collection<LookupElement> items = CompletionProviderUtils.createPsiItems(processor.getVariables());
-        result.addAllElements(CompletionProviderUtils.wrapInGroup(CompletionGrouping.NormalVar.ordinal(), items));
+        Collection<LookupElement> items = CompletionProviderUtils.createFromPsiItems(processor.getVariables(), BashIcons.VAR_ICON, CompletionGrouping.NormalVar.ordinal());
+        result.addAllElements(items);
 
         return items.size();
     }
 
     private void addGlobalVariables(CompletionResultSet result, Project project) {
         if (BashProjectSettings.storedSettings(project).isAutcompleteGlobalVars()) {
-            Collection<LookupElement> globalVars = CompletionProviderUtils.createItems(BashProjectSettings.storedSettings(project).getGlobalVariables(), BashIcons.GLOBAL_VAR_ICON);
-            result.addAllElements(CompletionProviderUtils.wrapInGroup(CompletionGrouping.GlobalVar.ordinal(), globalVars));
+            Collection<LookupElement> globalVars = CompletionProviderUtils.createItems(BashProjectSettings.storedSettings(project).getGlobalVariables(), BashIcons.GLOBAL_VAR_ICON, CompletionGrouping.GlobalVar.ordinal());
+            result.addAllElements(globalVars);
         }
     }
 
     private void addBuildInVariables(CompletionResultSet result, Project project) {
         if (BashProjectSettings.storedSettings(project).isAutocompleteBuiltinVars()) {
-            Collection<LookupElement> shellBuiltIns = CompletionProviderUtils.createItems(LanguageBuiltins.bashShellVars, BashIcons.BASH_VAR_ICON);
-            result.addAllElements(CompletionProviderUtils.wrapInGroup(CompletionGrouping.BuiltInVar.ordinal(), shellBuiltIns));
+            Collection<LookupElement> shellBuiltIns = CompletionProviderUtils.createItems(LanguageBuiltins.bashShellVars, BashIcons.BASH_VAR_ICON, CompletionGrouping.BuiltInVar.ordinal());
+            result.addAllElements(shellBuiltIns);
 
-            Collection<LookupElement> bashBuiltIns = CompletionProviderUtils.createItems(LanguageBuiltins.bourneShellVars, BashIcons.BOURNE_VAR_ICON);
-            result.addAllElements(CompletionProviderUtils.wrapInGroup(CompletionGrouping.BuiltInVar.ordinal(), bashBuiltIns));
+            Collection<LookupElement> bashBuiltIns = CompletionProviderUtils.createItems(LanguageBuiltins.bourneShellVars, BashIcons.BOURNE_VAR_ICON, CompletionGrouping.BuiltInVar.ordinal());
+            result.addAllElements(bashBuiltIns);
         }
     }
 }