+++ src/com/ansorgit/plugins/bash/util/CompletionUtil.java
@@ -15,8 +15,6 @@
 
 package com.ansorgit.plugins.bash.util;
 
-import com.google.common.base.Predicate;
-import com.google.common.base.Predicates;
 import com.google.common.collect.Lists;
 import com.intellij.openapi.util.SystemInfoRt;
 import org.jetbrains.annotations.NotNull;
@@ -26,12 +24,13 @@ import java.io.FileFilter;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.function.Predicate;
 
 /**
  * Class to help with file path completions.
  * It can provide the possible matches for absolute and relative paths.
  */
-public class CompletionUtil {
+public final class CompletionUtil {
     private CompletionUtil() {
     }
 
@@ -73,7 +72,7 @@ public class CompletionUtil {
         List<String> result = Lists.newLinkedList();
 
         for (File fileCandidate : collectFiles(basePath, matchPrefix)) {
-            if (!accept.apply(fileCandidate)) {
+            if (!accept.test(fileCandidate)) {
                 continue;
             }
 
@@ -104,7 +103,7 @@ public class CompletionUtil {
 
         String bashBaseDir = OSUtil.toBashCompatible(baseDir);
 
-        for (String path : completeAbsolutePath(baseDir + File.separator + relativePath, Predicates.<File>alwaysTrue())) {
+        for (String path : completeAbsolutePath(baseDir + File.separator + relativePath, file -> true)) {
             if (path.startsWith(bashBaseDir)) {
                 result.add(shownBaseDir + path.substring(bashBaseDir.length()));
             }